Extract RecomputePnpButton; report visible/hidden tags on failure
Move the PnP recompute button into its own RecomputePnpButton component.
The recompute now grabs a fresh still frame and re-detects the aruco
tags before re-solving PnP. When not all tags are visible it fails with
a toast reporting how many aruco tags are visible vs hidden.
